refactor: centralize matching file code
